version reconciling comments of the second OAB review
re correspondence of river positioning system with OGC standard model on linear referencing (OGC AS Topic 19): description of how features basically can be located along a flowpath, and in detail features realizing an outfall, features not realizing but referencing an outfall and features referencing a feature not realizing a catchment outfall, and how features can be located in the flowpath network. To this new figures 14 - 17 are included illustrating the concept of indirect position along the one-dimensional flowpath. The location of features in the hydrologic flowpath network is then illustrated using the still existing catchment graphics (now numbered as figs 18-20). This figures also show by means of newly included ‘tagged values’ how Indirect position concepts may  correspond to a 19184: Position expression. Table 1 still provides a (revised) ‘descriptive mapping’ including references to the normative part.